Move EventSchedule as an Event subcomponent
