--- content/webapp/views/components/EventSchedule/README.mdx
@@ -1,5 +1,0 @@
-## Purpose
-To show a list of events over a constrained period of time.
-
-
-[Edit this Readme on GitHub](https://github.com/wellcomecollection/wellcomecollection.org/edit/main/content/webapp/components/EventSchedule/README.mdx)
